Allow blank referers, fixed imagefap and xhamster
User complained about not being able to download zips from his android
device; this was due to the anti-hotlinking script in .htaccess.

imagefap would not download the first image in an album if the user's
avatar was the default 'hand.' fixed.

xhamster updated their site design. updated script to handle the
changes.